+++ packages/modal/src/vue/composables/useWeb3AuthConnect.ts
@@ -1,4 +1,5 @@
 import { Connection, log, LoginParamMap, WALLET_CONNECTOR_TYPE, WalletInitializationError, Web3AuthError } from "@web3auth/no-modal";
+import { useWeb3AuthConnect as useWeb3AuthNoModalConnect } from "@web3auth/no-modal/vue";
 import { Ref, ref, watch } from "vue";
 
 import { useWeb3AuthInner } from "./useWeb3AuthInner";
@@ -13,56 +14,52 @@ export interface IUseWeb3AuthConnect {
 }
 
 export const useWeb3AuthConnect = (): IUseWeb3AuthConnect => {
-  const { web3Auth, isConnected } = useWeb3AuthInner();
+  const { web3Auth } = useWeb3AuthInner();
+  // Web3AuthModal extends the NoModal class, so we can use the same composable
+  const { isConnected, loading: sharedLoading, error: sharedError, connectorName, connect: sharedConnectTo } = useWeb3AuthNoModalConnect();
   const loading = ref(false);
   const error = ref<Web3AuthError | null>(null);
-  const connectorName = ref<WALLET_CONNECTOR_TYPE | null>(null);
+  const connectLoading = ref(false);
+  const connectError = ref<Web3AuthError | null>(null);
+  const activeFlow = ref<"connect" | "connectTo" | null>(null);
 
   watch(
-    isConnected,
-    (newVal) => {
-      if (!newVal && connectorName.value) {
-        connectorName.value = null;
-      }
-      if (newVal && !connectorName.value) {
-        connectorName.value = web3Auth.value?.connectedConnectorName;
-      }
+    [connectLoading, sharedLoading],
+    ([newConnectLoading, newSharedLoading]) => {
+      loading.value = newConnectLoading || newSharedLoading;
+    },
+    { immediate: true }
+  );
+
+  watch(
+    [activeFlow, connectError, sharedError],
+    ([newActiveFlow, newConnectError, newSharedError]) => {
+      error.value = newActiveFlow === "connect" ? newConnectError : newSharedError;
     },
     { immediate: true }
   );
 
   const connect = async () => {
     try {
+      activeFlow.value = "connect";
       if (!web3Auth.value) throw WalletInitializationError.notReady();
-      error.value = null;
-      loading.value = true;
+      connectError.value = null;
+      connectLoading.value = true;
       const localProvider = await web3Auth.value.connect();
-      connectorName.value = web3Auth.value.connectedConnectorName;
       return localProvider;
     } catch (err) {
       log.error("Error connecting", err);
-      error.value = err as Web3AuthError;
+      connectError.value = err as Web3AuthError;
       return null;
     } finally {
-      loading.value = false;
+      connectLoading.value = false;
     }
   };
 
   const connectTo = async <T extends WALLET_CONNECTOR_TYPE>(connectorType: T, loginParams?: LoginParamMap[T]) => {
-    try {
-      if (!web3Auth.value) throw WalletInitializationError.notReady();
-      error.value = null;
-      loading.value = true;
-      const localProvider = await web3Auth.value.connectTo(connectorType, loginParams);
-      connectorName.value = web3Auth.value.connectedConnectorName;
-      return localProvider;
-    } catch (err) {
-      log.error("Error connecting to", err);
-      error.value = err as Web3AuthError;
-      return null;
-    } finally {
-      loading.value = false;
-    }
+    activeFlow.value = "connectTo";
+    connectError.value = null;
+    return sharedConnectTo(connectorType, loginParams);
   };
 
   return {
