+++ packages/no-modal/src/vue/Web3AuthProvider.ts
@@ -1,223 +1,22 @@
-import { defineComponent, h, PropType, provide, ref, shallowRef, watch } from "vue";
+import { defineComponent, h, PropType, provide } from "vue";
 
-import {
-  ANALYTICS_INTEGRATION_TYPE,
-  type ChainNamespaceType,
-  type Connection,
-  CONNECTOR_EVENTS,
-  CONNECTOR_STATUS,
-  type CONNECTOR_STATUS_TYPE,
-  log,
-  WalletInitializationError,
-  Web3AuthContextKey,
-} from "../base";
+import { Web3AuthContextKey } from "../base";
 import { Web3AuthNoModal } from "../noModal";
 import { IWeb3AuthInnerContext, Web3AuthContextConfig } from "./interfaces";
+import { useWeb3AuthInnerContextValue } from "./useWeb3AuthInnerContextValue";
 import { WalletServicesInnerProvider } from "./WalletServicesInnerProvider";
 
 export const Web3AuthProvider = defineComponent({
   name: "Web3AuthProvider",
   props: { config: { type: Object as PropType<Web3AuthContextConfig>, required: true } },
   setup(props) {
-    const web3Auth = shallowRef<Web3AuthNoModal | null>(null);
-    const connection = ref<Connection | null>(null);
-    const isMFAEnabled = ref(false);
-    const status = ref<CONNECTOR_STATUS_TYPE | null>(null);
-    const chainId = ref<string | null>(null);
-    const chainNamespace = ref<ChainNamespaceType | null>(null);
-
-    const isInitializing = ref(false);
-    const initError = ref<Error | null>(null);
-    const isInitialized = ref(false);
-
-    const isConnected = ref(false);
-    const isAuthorized = ref(false);
-
-    const getPlugin = (name: string) => {
-      if (!web3Auth.value) throw WalletInitializationError.notReady();
-      return web3Auth.value.getPlugin(name);
-    };
-
-    const setIsMFAEnabled = (isMfaEnabled: boolean) => {
-      isMFAEnabled.value = isMfaEnabled;
-    };
-
-    watch(
-      () => props.config,
-      (newConfig, _, onInvalidate) => {
-        const resetHookState = () => {
-          connection.value = null;
-          isMFAEnabled.value = false;
-          isConnected.value = false;
-          isAuthorized.value = false;
-          status.value = null;
-          chainId.value = null;
-          chainNamespace.value = null;
-        };
-
-        onInvalidate(() => {
-          if (web3Auth.value) {
-            web3Auth.value.cleanup();
-          }
-        });
-
-        resetHookState();
-        const { web3AuthOptions } = newConfig;
-        const web3AuthInstance = new Web3AuthNoModal(web3AuthOptions);
-        web3AuthInstance.setAnalyticsProperties({
-          integration_type: ANALYTICS_INTEGRATION_TYPE.VUE_COMPOSABLES,
-        });
-        web3Auth.value = web3AuthInstance;
-      },
-      { immediate: true }
-    );
-
-    watch(
-      web3Auth,
-      async (newWeb3Auth, _, onInvalidate) => {
-        const controller = new AbortController();
-
-        // Invalidate the controller here before calling any async methods.
-        onInvalidate(() => {
-          controller.abort();
-        });
-
-        if (newWeb3Auth) {
-          try {
-            initError.value = null;
-            isInitializing.value = true;
-            await newWeb3Auth.init({ signal: controller.signal });
-          } catch (error) {
-            log.error("Error initializing web3auth", error);
-            initError.value = error as Error;
-          } finally {
-            isInitializing.value = false;
-          }
-        }
-      },
-      { immediate: true }
-    );
-
-    watch(
-      web3Auth,
-      (newWeb3Auth, prevWeb3Auth) => {
-        const notReadyListener = () => {
-          status.value = web3Auth.value!.status;
-        };
-        const readyListener = () => {
-          status.value = web3Auth.value!.status;
-          isInitialized.value = true;
-        };
-
-        const connectedListener = () => {
-          status.value = web3Auth.value!.status;
-          // we do this because of rehydration issues. status connected is fired first but web3auth sdk is not ready yet.
-          if (web3Auth.value!.status === CONNECTOR_STATUS.CONNECTED) {
-            if (!isInitialized.value) isInitialized.value = true;
-            isConnected.value = true;
-            connection.value = newWeb3Auth.connection;
-            chainId.value = web3Auth.value!.currentChainId;
-            chainNamespace.value = web3Auth.value!.currentChain?.chainNamespace ?? null;
-          }
-        };
-
-        const authorizedListener = () => {
-          status.value = web3Auth.value!.status;
-          if (web3Auth.value!.status === CONNECTOR_STATUS.AUTHORIZED) {
-            isAuthorized.value = true;
-            isConnected.value = true;
-          }
-        };
-
-        const disconnectedListener = () => {
-          status.value = web3Auth.value!.status;
-          isConnected.value = false;
-          connection.value = null;
-          isMFAEnabled.value = false;
-          isAuthorized.value = false;
-        };
-
-        const connectingListener = () => {
-          status.value = web3Auth.value!.status;
-        };
-
-        const errorListener = () => {
-          status.value = web3Auth.value!.status;
-        };
-
-        const mfaEnabledListener = () => {
-          isMFAEnabled.value = true;
-        };
-
-        // unregister previous listeners
-        if (prevWeb3Auth && newWeb3Auth !== prevWeb3Auth) {
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.NOT_READY, notReadyListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.READY, readyListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.CONNECTED, connectedListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.AUTHORIZED, authorizedListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.DISCONNECTED, disconnectedListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.CONNECTING, connectingListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.ERRORED, errorListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.REHYDRATION_ERROR, errorListener);
-          prevWeb3Auth.removeListener(CONNECTOR_EVENTS.MFA_ENABLED, mfaEnabledListener);
-        }
-
-        if (newWeb3Auth && newWeb3Auth !== prevWeb3Auth) {
-          status.value = newWeb3Auth.status;
-          // web3Auth is initialized here.
-          newWeb3Auth.on(CONNECTOR_EVENTS.NOT_READY, notReadyListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.READY, readyListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.CONNECTED, connectedListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.AUTHORIZED, authorizedListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.DISCONNECTED, disconnectedListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.CONNECTING, connectingListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.ERRORED, errorListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.REHYDRATION_ERROR, errorListener);
-          newWeb3Auth.on(CONNECTOR_EVENTS.MFA_ENABLED, mfaEnabledListener);
-        }
-      },
-      { immediate: true }
-    );
-
-    // Listen for chain changes on the provider (derived from connection)
-    watch(
-      connection,
-      (newConnection, prevConnection) => {
-        const handleChainChange = (newChainId: string) => {
-          chainId.value = newChainId;
-          chainNamespace.value = web3Auth.value?.currentChain?.chainNamespace ?? null;
-        };
-
-        const prevProvider = prevConnection?.ethereumProvider ?? null;
-        const newProvider = newConnection?.ethereumProvider ?? null;
-
-        // unregister previous listener
-        if (prevProvider && newProvider !== prevProvider) {
-          prevProvider.removeListener("chainChanged", handleChainChange);
-        }
-
-        if (newProvider && newProvider !== prevProvider) {
-          newProvider.on("chainChanged", handleChainChange);
-        }
-      },
-      { immediate: true }
-    );
-
-    provide<IWeb3AuthInnerContext>(Web3AuthContextKey, {
-      web3Auth,
-      isConnected,
-      isInitialized,
-      connection,
-      status,
-      isInitializing,
-      initError,
-      isMFAEnabled,
-      chainId,
-      chainNamespace,
-      getPlugin,
-      setIsMFAEnabled,
-      isAuthorized,
+    const value = useWeb3AuthInnerContextValue({
+      Web3AuthConstructor: Web3AuthNoModal,
+      watchSource: () => props.config,
+      getWeb3AuthOptions: (config) => config.web3AuthOptions,
     });
+
+    provide<IWeb3AuthInnerContext>(Web3AuthContextKey, value);
   },
   render() {
     return h(WalletServicesInnerProvider, {}, this.$slots.default ?? "");
