+++ packages/no-modal/src/vue/index.ts
@@ -2,4 +2,5 @@ export * from "./composables";
 export * from "./context";
 export * from "./interfaces";
 export { useWalletServicesInnerContextValue } from "./useWalletServicesInnerContextValue";
+export { useWeb3AuthInnerContextValue } from "./useWeb3AuthInnerContextValue";
 export * from "./Web3AuthProvider";
